refactor BatchNode hooks and use NodeContextMenu
